refactor: split nanogpt catalog modules
