+++ package.json
@@ -30,6 +30,8 @@
     "runtime/provider-pricing.ts",
     "runtime/routing.ts",
     "runtime/usage.ts",
+    "catalog/models-json-snapshot.ts",
+    "catalog/build-provider.ts",
     "provider-catalog.ts",
     "provider-discovery.ts",
     "nanogpt-errors.ts",
